+++ src/browser/simulatedstore/showSimulatedPurchaseAlert.ts
